Executable to extract convolutionable samples from a given text
The tool receives how much context to add around a
character. `text-to-char-context N` ouputs all windows like

  x x x C x x x

Where x are the context chars (N of them to the left and N to the right) and C
is the char to predict.

Initial and last windows are padded with NULL bytes (0x00)