Make sandboxing arguments explicit
For some reason, in the current version of nix I get `{}` as argument instead of
`false` when the argument is not passed (smells like a bug).

All in all, explicit is usually better, so let's go with this. Although it makes
it more difficult for the haskell project to be sandboxes with a generic script