Slightly less crufty implementation
Move that std::array doesn't move, so removing those.

Using arrays may seem as lots of unnecessary copies, but tried with vector and
there are a lot more allocs.

Anyway, this runs in 3 ms in my laptop, so all these optimisations are pure
experimentation, we don't need them for an 8x8 board