Significantly faster solution
We are now initialising the Bfs object twice, rather than twice the number of
connected components. Taking into account that Bfs initialises in O(V) (being V
the number of vertices), this reduces the worst case from O(V^2) to O(V) and
solves the current nemesis in subsecons time.