Add advent 2019 to the new haskell packages structure
And reworked haskell-pkg a bit so that it only needs haskellPackages as input,
as we don't need to add dependencies apart from it now.

Note that there is a bit of a patch, since `callPackage` from haskellPackages
doesn't seem to provide access to the override haskellPackages, but to
super.haskellPackages.