Slightly better error handling
Fail on any error other than ExitCode when running sexp. Before we were ignoring
any error provided that the error output was empty, but that was the case even
when sexp could not run at all.

Cleaner messages and more aggressive panics, as I really don't want to have 90%
of my code be error handling code.